drm/i915: Correcting delay's from u8 to u16
backlight_off delay, panel_on delay, panel off delay etc.
datatype should be u16 as the values can go beyond 255.

Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-12646
Reviewed-on: https://android.intel.com/#/c/307845/
Reviewed-on: https://android.intel.com:443/307845
Reviewed-on: https://android.intel.com:443/309484

+++ drivers/gpu/drm/i915/intel_dsi.h
@@ -196,14 +196,14 @@ struct intel_dsi {
 	u16 port;
 
 	/* all delays in ms */
-	u8 backlight_off_delay;
-	u8 backlight_on_delay;
-	u8 panel_on_delay;
-	u8 panel_off_delay;
-	u8 panel_pwr_cycle_delay;
+	u16 backlight_off_delay;
+	u16 backlight_on_delay;
+	u16 panel_on_delay;
+	u16 panel_off_delay;
+	u16 panel_pwr_cycle_delay;
 
 	bool send_shutdown;
-	u8 shutdown_pkt_delay; /*in ms*/
+	u16 shutdown_pkt_delay; /*in ms*/
 	enum panel_fitter pfit;
 };
